feat: add `time_walk_env_binary` and refactor `time_walk_custom` with shared base logic

+++ src/parsers/environment_params.h
@@ -12,7 +12,7 @@ EnvironmentInfluenceGrid *parse_kernel_params(const char *csv_data, const DateTi
                                               const Dimensions3D *dims);
 
 KernelParamsYXT *
-get_kernels_environment_grid(int T, const TerrainMap *terrain, const EnvironmentInfluenceGrid *grid,
+get_kernels_environment_grid(size_t T, const TerrainMap *terrain, const EnvironmentInfluenceGrid *grid,
                              const KernelParametersMapping *kernels_mapping, float environment_weight);
 
 void free_environment_influence_grid(EnvironmentInfluenceGrid *grid);
